requirements.txt
APScheduler==2.1.2
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
cryptography==2.6.1
Django==2.2
django-ckeditor==5.7.0
django-debug-toolbar==1.11
django-js-asset==1.2.2
django-mysql==3.0.0.post1
django-qiniu-storage==2.3.1
idna==2.8
isort==4.3.17
jieba==0.39
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
orjson==2.0.5
Pillow==6.0.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
qiniu==7.2.2
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
urllib3==1.24.2
wrapt==1.11.1
xmltodict==0.11.0
PyMySQL==0.9.3